+++ .env.development
@@ -16,7 +16,7 @@ DHIS2_PASSWORD=
 DHIS2_URL=
 DHIS2_USERNAME=
 DHIS2_VERSION="2.28" # This is a hack because the current dhis2 gem only works against version 2.28 of the DHIS API
-DHIS2_DATA_ELEMENTS_FILE="config/data/dhis2/local.yml"
+DHIS2_DATA_ELEMENTS_FILE="config/data/dhis2/sandbox.yml"
 EMAIL_SUBJECT_PREFIX="[Simple DEVELOPMENT]"
 EXOTEL_API_RATE_LIMIT_PER_MINUTE=250
 EXOTEL_CALL_DETAILS_JOB_RETRY_SECONDS=600
