--- app/exporters/bangladesh_disaggregated_dhis2_exporter.rb
@@ -1,12 +1,0 @@
-class BangladeshDisaggregatedDhis2Exporter
-  TOTAL_MONTHS = 24
-
-  def self.export
-    FacilityBusinessIdentifier.dhis2_org_unit_id.each do |facility_identifier|
-      Dhis2::BangladeshDisaggregatedExporterJob.perform_async(
-        facility_identifier.id,
-        TOTAL_MONTHS
-      )
-    end
-  end
-end
