+++ config/schedule.rb
@@ -106,7 +106,7 @@ end
 
 every 1.month, at: local("04:00 am"), roles: [:cron] do
   if Flipper.enabled?(:dhis2_export)
-    rake "dhis2:export"
+    rake "dhis2:bangladesh_export"
   end
 end
